refactored for readability
